+++ CHANGELOG.md
@@ -29,6 +29,7 @@
   not-yet-implented expression.
 - Holes can be used in type annotations to specify part of a type, leaving the
   rest for inference.
+- The incorrect arity error now prints any missing labelled arguments.
 
 ## v0.10.1 - 2020-07-15